pyproject.toml
[project]
name = "argus-osint"
version = "0.1.0"
description = "Multi-agent OSINT platform for cross-platform person discovery and verification"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Internet",
    "Typing :: Typed",
]
dependencies = [
    "aiohttp",
    "aiohttp-retry",
    "beautifulsoup4",
    "lxml",
    "pydantic>=2.0",
    "click",
    "rich",
    "imagehash",
    "scikit-learn",
    "jellyfish",
    "networkx",
    "python-dateutil",
    "Pillow",
    "aiosqlite",
]

[project.urls]
Homepage = "https://github.com/TomOst-Sec/J-R"
Repository = "https://github.com/TomOst-Sec/J-R"

[project.optional-dependencies]
playwright = ["camoufox[geoip]"]
face = ["face-recognition", "dlib"]
llm = ["openai", "tiktoken"]
api = ["fastapi", "uvicorn", "websockets"]
mcp = ["mcp"]
encrypted = ["sqlcipher3"]
intel = ["shodan", "aiodns", "python-whois", "phonenumbers"]
all = [
    "camoufox[geoip]",
    "openai", "tiktoken",
    "fastapi", "uvicorn", "websockets",
    "mcp",
    "shodan", "aiodns", "python-whois", "phonenumbers",
]

[project.scripts]
argus = "argus.cli:main"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "aioresponses",
    "ruff",
    "mypy",
    "faker",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/argus"]
